// cpu-app/ClientApp/src/app/core/api/models/eligibleExpenseItemDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */

export interface EligibleExpenseItemDto {
  /** @nullable */
  vsd_EligibleExpenseItemId?: string | null;
  /** @nullable */
  vsd_Name?: string | null;
  /** @nullable */
  vsd_ProgramExpenseType?: number | null;
  /** @nullable */
  fortuneCookieType?: string | null;
  /** @nullable */
  fortuneCookieEtag?: string | null;
}