// cpu-app/ClientApp/src/app/core/api/models/programSurplusDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CPU Portal API
 * API documentation for CPU Portal
 * OpenAPI spec version: v1
 */
import type { ContractDto } from './contractDto';
import type { ProgramDto } from './programDto';
import type { OrganizationDto } from './organizationDto';
import type { SurplusPlanDto } from './surplusPlanDto';
import type { SurplusLineItemDto } from './surplusLineItemDto';
import type { EligibleExpenseItemDto } from './eligibleExpenseItemDto';

export interface ProgramSurplusDto {
  isSuccess?: boolean;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  businessbceid?: string | null;
  /** @nullable */
  userbceid?: string | null;
  contract?: ContractDto;
  program?: ProgramDto;
  organization?: OrganizationDto;
  surplusPlan?: SurplusPlanDto;
  /** @nullable */
  surplusPlanLineItems?: SurplusLineItemDto[] | null;
  /** @nullable */
  eligibleExpenseItemCollection?: EligibleExpenseItemDto[] | null;
}